css-text-3/overflow-wrap/overflow-wrap-break-word-fit-content-001.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>CSS Text Test: overflow-wrap: break-word with fit-content</title>
<link rel="author" title="Koji Ishii" href="kojiishi@gmail.com">
<link rel="help" title="6.2. Overflow Wrapping: the 'word-wrap'/'overflow-wrap' property" href="http://www.w3.org/TR/css-text-3/#overflow-wrap">
<meta name="flags" content="dom">
<meta name="assert" content="The 'overflow-wrap: break-word' must not break surrogate pairs">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<style>
.fit {
  display: inline-block;
  word-wrap: break-word;
  font-size: 20px;
  line-height: 1;
}
</style>
<body>
<div id="log"></div>
<p classs="instructions">Tests pass if all lines do not wrap.
<ol id="container">
  <li><div class="fit">&#x1D70B;</div>
  <li><div class="fit">ğœ‹</div>
</ol>
<script>
var tests = document.querySelectorAll(".fit");
for (var i = 0; i < tests.length; i++) {
  var element = tests[i];
  test(function () {
    var rects = element.getClientRects();
    assert_equals(1, rects.length);
    assert_equals(rects[0].height, 20);
  }, (i + 1).toString());
}
</script>
</body>
